Moving from using a surrogate key to a composite key.
Schema: Prior to this update, my PA was using a surrogate key (but I was bad and forgot to update the repo copy). Obviously, we need the "entity id", "entity type" and url to be part of the composite key. But MySQL on PA by default won't allow keys larger than 767. So decided to take a SHA512 hash of the url and use that for a key (length 128, fyi). We still retain the full URL (so a human can go visit it later).

Decided to move the URL across to a TEXT field, simply because there's no actual fully accepted upper limit on URL length. I think 2000 was a de-facto length used, but why limit ourselves?

Also updated the timestamp to be a BIGINT so we can use 64bit unix epoch timestamps.

--

Flask app: no change to the API, just hashing the url when it comes in. We also insert or update instead of mindlessly duplicating entries.